Last commit 'til next week.  Just clearing out my local changes before taking a (short) break from code and admin.
 - Use stdbool.h when HAVE_STDBOOL_H is defined.
 - Use uint8_t for byte when available
 - Fix signedness comparisons (util.c, object1.c)
 - Only include <limits.h> once (Kenneth Boyd)
 - Split timeout decrements into its own function, and remove duplicated code (thanks to Kenneth Boyd for pointing this out)
 - Fix warnings in ui.h
 - Fix dependencies for Makefile.inc




git-svn-id: svn://rephial.org/angband/trunk@87 98ddb578-d4ae-11db-a93f-5346e46f259c

+++ src/object1.c
@@ -169,7 +169,7 @@ void flavor_init(void)
 		bool okay = TRUE;
 
 		wordlen = make_word(RANDNAME_SCROLL, 2, 8, end, 24);
-		while (titlelen + wordlen < (sizeof(scroll_adj[0]) - 1))
+		while (titlelen + wordlen < (int)(sizeof(scroll_adj[0]) - 1))
 		{
 			end[wordlen] = ' ';
 			titlelen += wordlen + 1;
